# Speed up `CoreCaskTap#cask_files_by_name`